♻️ refactor: 공지 작성 폼을 공통 컴포넌트로 분리하고 라우트에 적용
