+++ src/features/notice/index.ts
@@ -3,4 +3,5 @@ export { type Chapter, CHAPTERS, ChapterSelector } from "./ui/ChapterSelector"
 export { CheckBox } from "./ui/CheckBox"
 export { NoticeCard } from "./ui/NoticeCard"
 export { NoticeCardList, type NoticeItem } from "./ui/NoticeCardList"
+export { NoticePublishForm } from "./ui/NoticePublishForm"
 export { NoticeSubmitButton } from "./ui/NoticeSubmitButton"
